Refactor download handler into fixture (so it can be shared between tests).
