+++ test/download_handler/exception_test.py
@@ -2,32 +2,14 @@ from contextlib import nullcontext as does_not_raise
 from typing import Any
 
 import pytest
+from _fixture.download_handler import get as get_download_handler
 from _helper import directory, file
 from _helper.timeout import reset_to_not_timed_out
 from py.path import local
 
-from browserist import Browser, BrowserSettings, factory
-from browserist.constant import idle_timeout
+from browserist import Browser, BrowserSettings
 from browserist.exception.download import (DownloadHandlerMultipleFinalFilesError,
                                            DownloadHandlerMultipleTemporaryFilesError)
-from browserist.model.download.handler import DownloadHandler
-
-
-def get_download_handler(browser: Browser, download_dir_entries_before_download: list[str]) -> DownloadHandler:
-    def ensure_uses_temporary_file_is_true(download_handler: DownloadHandler) -> DownloadHandler:
-        property(
-            fget=lambda value: True,
-            fset=setattr(download_handler, "_uses_temporary_file.setter", lambda: True),
-            fdel=None
-        )
-        return download_handler
-
-    download_handler = factory.get.download_handler(
-        browser_driver=browser._browser_driver,
-        download_dir_entries_before_download=download_dir_entries_before_download,
-        idle_download_timeout=idle_timeout.VERY_SHORT,
-    )
-    return ensure_uses_temporary_file_is_true(download_handler)
 
 
 @pytest.mark.parametrize("number_of_files, expectation", [
